requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
chardet==3.0.4
distro==1.3.0
django==2.1.3
dominate==2.3.4
flask==1.0.2
flask-bootstrap==3.3.7.1
flatdict==3.1.0
idna==2.7
itsdangerous==1.1.0
jinja2==2.10
more-itertools==4.3.0
numpy==1.15.4
pandas==0.23.4
pathlib==1.0.1
pluggy==0.8.0
prometheus-client==0.4.2
protobuf==3.6.1
protobuf-to-dict==0.1.0
py==1.7.0
pypdf2==1.26.0
pytest==4.0.0
pytest-runner==4.2
python-dateutil==2.7.5
python3-protobuf==2.5.0
pyyaml==3.13
requests==2.20.1
six==1.11.0
tabula-py==1.3.1
urllib3==1.24.1
visitor==0.1.3
wget==3.2
dpath==1.4.2
robotframework-python3==2.9
